requirements.txt
paho-mqtt == 2.0.0
g29py == 0.0.8
hidapi == 0.14.0